% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring_cox_est.R
\name{censoring_cox_est}
\alias{censoring_cox_est}
\title{Compute survival function and other related results of censoring based on Cox proportional hazards model}
\usage{
censoring_cox_est(Data, SurvTime, Status, covariateC)
}
\arguments{
\item{Data}{A data frame includes survival time, censoring indicator and related covariates for censoring}

\item{SurvTime}{name of survival time for fatal endpoint (e.g., death) in Data}

\item{Status}{name of status for fatal endpoint in Data}

\item{covariateC}{a vector of covariates used in Cox model for censoring}
}
\value{
A list with the following elements:
\describe{
\item{Kij}{a matrix for survival functions of censoring, and the element in (i,j) refers to the survival function of censoring for ith subject at observed time t_j}
\item{Omega}{a variance-covariance matrix for coefficients in Cox model for cesoring}
\item{G}{value of G for subject i at its observed time t_i, a n times pc (dimension of covariateC) matrix}
\item{Psi}{value of martingle integral for subject i, a n times pc (dimension of covariateC) matrix}
}
}
\description{
This function can obtain survival function of censoring, G, U and Omega results
based on Cox model. Their concrete formulas are shown in Section 3 of paper (Cao et al., 2026).
}
\references{
Cao Z., Fang X. and Li F. Generalized win fraction regression for composite survival endpoints.
under review. 2026;0(0):1-35.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
