% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwfm.R
\name{gwfm}
\alias{gwfm}
\title{A generalized win fraction regression model for composite survival data}
\usage{
gwfm(
  formula,
  data,
  SurvTime,
  Status,
  covariateC = NULL,
  priority = 1:2,
  L = Inf,
  t_val = 0.01,
  IPCW = TRUE,
  start = NULL,
  control = list(),
  link = c("logit", "probit", "identity"),
  model = c("difference", "marginal", "customized"),
  compare = if (model == "marginal") "all" else "unique",
  na.action = getOption("na.action"),
  keep.data = FALSE,
  ...
)
}
\arguments{
\item{formula}{An object of class formula (or one that can be coerced to that class): A symbolic description of the model to be fitted. The details of model specification are given under 'Details'.}

\item{data}{a data frame including id, multiple endpoints (e.g., y1, y2), corresponding censoring status (e.g., delta1, delta2) and covariates}

\item{SurvTime}{name of survival time for fatal endpoint (e.g., death)}

\item{Status}{name of status for fatal endpoint}

\item{covariateC}{covariates used in Cox regression model for censoring when IPCW method is chosen.
If IPCW = FALSE, then this setting is useless (default NULL)}

\item{priority}{composite endpoints ranked in descending order of importance (e.g., priority = c(1,2) means y1 is more important than y2)}

\item{L}{pre-specified time horizon to construct restriction event endpoints (default Inf)}

\item{t_val}{truncation value used in IPCW weight (default 0.01)}

\item{IPCW}{using IPCW adjustment or not (default TRUE)}

\item{start}{start value for fitting a gwfm model for composite survival data (default NULL)}

\item{control}{control settings for \code{\link[nleqslv]{nleqslv}} when fitting gwfm model for composite survival data (default list())}

\item{link}{a character vector with a single value that determines the used link function. Possible values are "logit", "probit" and "identity". The default is "logit".}

\item{model}{a single character value with possible values "difference" (the default), "marginal", or "customized". If the formula indicates a customized model (by the
use of L() or R() in \code{\link{pim}} package), this parameter is set automatically to "customized".}

\item{compare}{a character vector with a single value that describes how the model compares observations. It can take the values "unique" or "all". Alternatively you can pass
a matrix with two columns. Each row represents the row numbers in the original data frame that should be compared to each other. See Details.}

\item{na.action}{the name of a function which indicates what should happen when the data contains NAs. The default is set by the na.action setting of options, and is na.fail when unset.}

\item{keep.data}{a logical value indicating whether the model matrix should be saved in the object. Defaults to FALSE. See Details}

\item{...}{extra parameters sent to \code{\link{gwfm_fit}}}
}
\value{
A list with the following elements
\describe{
\item{summary}{point estimation, SE and P value of parameter coefficients in gwfm}
\item{P_tie}{tied probability}
\item{coef}{point estimation of parameter coefficients in gwfm}
\item{vcov}{a numeric matrix with the variance-covariance matrix for the coefficients}
\item{fitted}{a numeric vector with the raw fitted values}
\item{flag}{a flag indicating nleqslv is converged or not, its value is equal to termcd of nleqslv}
\item{penv}{a gwfm environment from a model or formula}
\item{link}{link function used in gwfm}
\item{estim}{a list with two components named \code{coef} and \code{vcov}
containing information on the used estimators for both.}
\item{model.matrix}{a design matrix for a gwfm model}
\item{response}{pseudo-observed response for win function. 1 for win result, 0 for loss or tied comparison result}
\item{keep.data}{a logical value indicating whether the model matrix should be saved in the object}
\item{model}{the chosen model in a gwfm}
}
}
\description{
This function fits a generalized win fraction regression model (gwfm), also known as GWFM.
It can be used to fit many different flavours of models that can be reformulated as a gwfm. The most
general models are implemented, but the flexible formula interface allows you to specify a wide
variety of different models, like the \code{\link{pim}} package
}
\details{
GWFMs are based on a set of pseudo-observations constructed from the comparison between a range
of possible combinations of 2 observations with composite survival endpoints. We call the set of pseudo observations poset in the
context of this package, like \code{\link{pim}} package.

By default, this poset takes every unique combination of 2 observations (compare = "unique"). You
can either use a character value, or use a matrix or list to identify the set of observation pairs that
have to be used as pseudo-observations. Note that the matrix and list should be either nameless,
or have the (col)names ’L’ and ’R’. If any other names are used, these are ignored and the first
column/element is considered to be ’L’. See also \code{\link{new.pim.poset}} in \code{\link{pim}} package.

It’s possible to store the model matrix and psuedo responses in the resulting object. By default this
is not done (keep.data = FALSE) as this is less burden on the memory and the gwfm object
contains all information to reconstruct both the model matrix and the pseudo responses. If either
the model matrix or the pseudo responses are needed for further calculations, setting keep.data to
TRUE might reduce calculation time for these further calculations.
}
\examples{
# example 1: using gwfm with logit link
library(gumbel)
library(MASS)
library(pim)
library(copula)
library(nleqslv)
library(survival)
library(mvtnorm)
n = 200  
beta = c(-0.3,0.3) 
betac = c(-0.6,0.4) 
lambda_d = 0.2
lambda_h = 2
alpha = 2
set.seed(123456)
mydata1 = sim_gwfm_logit(n,beta,betac,lambda_d,lambda_h,alpha)

#no restriction event time, default L=Inf
res1 = gwfm(formula=y1~x1+x2,data=mydata1,SurvTime="y1",Status="delta1",
covariateC=c("x1","x2"),IPCW=FALSE, priority = 1:2, 
model = "difference",link = "logit")
print(res1$summary)

# restriction event time and Broyden method in nleqslv
res2 = gwfm(formula=y1~x1+x2,data=mydata1,SurvTime="y1",Status="delta1",
covariateC=c("x1","x2"), L = 1, IPCW=FALSE, priority = 1:2,
 model = "difference",link = "logit", method="Broyden")
print(res2$summary)
print(res2$WO)

# example 2: using gwfm with probit link
n = 200  
alpha = 1
betac = -1
u = 6
sig = 1
set.seed(123456)
mydata2 = sim_gwfm_probit(n,alpha,betac,u,sig)

# NO IPCW 
xtol = 1e-6; ftol = 1e-6
btol = 1e-3; maxit = 50
cntl = list(xtol=xtol, ftol=ftol, btol=btol, maxit=maxit)
res3 = gwfm(formula=y1~x,data=mydata2,SurvTime="y1",Status="delta1",
covariateC=c("x"), priority = 1:2, L = 7, IPCW = FALSE,
model = "difference",link = "probit",control = cntl)
print(res3$summary)

# IPCW for restriction event time L=7
res4 = gwfm(formula=y1~x,data=mydata2,SurvTime="y1",Status="delta1",
covariateC=c("x"),priority = 1:2, L = 7, model = "difference",
link = "probit",control = cntl)
print(res4$summary)

# IPCW for restriction event time L=8
res5 = gwfm(formula=y1~x,data=mydata2,SurvTime="y1",Status="delta1",
covariateC=c("x"), priority = 1:2, L = 8, model = "difference",
link = "probit", control = cntl)
print(res5$summary)

# example 3: using gwfm with identity link
n = 200  
beta_d = 10 
beta_h = 8
lambda_d = 0.15
lambda_h = 0.3 
betac = -5.7
set.seed(123456) 
mydata3 = sim_gwfm_identity(n,beta_d,beta_h,lambda_d,lambda_h,betac)

# NO IPCW 
res6 = gwfm(formula=y1~x,data=mydata3,SurvTime="y1",Status="delta1",
covariateC=c("z"), priority = 1:2, L = 15, IPCW = FALSE, 
model = "marginal",link = "identity",control = cntl)
print(res6$summary)

# IPCW for restriction time L=15
res7 = gwfm(formula=y1~x,data=mydata3,SurvTime="y1",Status="delta1",
covariateC=c("z"), priority = 1:2, L = 15, model = "marginal",
link = "identity",control = cntl)
print(res7$summary)

# IPCW for restriction time L=12
res8 = gwfm(formula=y1~x,data=mydata3,SurvTime="y1",Status="delta1",
covariateC=c("z"), priority = 1:2, L = 12, model = "marginal",
link = "identity",control = cntl)
print(res8$summary)

}
\references{
Cao Z., Fang X. and Li F. Generalized win fraction regression for composite survival endpoints.
under review. 2026;0(0):1-35.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
