% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwfm_fit.R
\name{gwfm_fit}
\alias{gwfm_fit}
\title{Fitter function for a generalized win fraction regression model for composite outcomes}
\usage{
gwfm_fit(
  x,
  y,
  link = "logit",
  start = rep(0, ncol(x)),
  control = list(),
  method = "Newton",
  vcov.estim = "sandwich.vcov",
  weight_est = NULL,
  weight_var = NULL,
  penv,
  ...
)
}
\arguments{
\item{x}{a model matrix with as many rows as \code{y}.}

\item{y}{a vector with the pseudo-responses}

\item{link}{a character vector with a link function}

\item{start}{A numeric vector with the starting values for the fitting algorithm, if required.}

\item{control}{conditions for convergence of \code{\link[nleqslv]{nleqslv}}}

\item{method}{"Newton" (default) and "Broyden" can be chosen in \code{\link[nleqslv]{nleqslv}}}

\item{vcov.estim}{a function to determine the variance-covariance matrix.
Possibilities are \code{\link{sandwich.vcov}} and \code{link{score.vcov}}.
Defaults to \code{sandwich.vcov}}

\item{weight_est}{weights used in parameter estimation}

\item{weight_var}{weights used in variance estimation}

\item{penv}{An environment}

\item{...}{Further arguments that need to be passed to the estimation function. The most
relevant is construct, allowing you to write your own score function for the
numerical optimization. See also estimators}
}
\value{
A list with the following elements
\describe{
\item{coefficients}{a numeric vector with the coefficients}
\item{fitted}{a numeric vector with the raw fitted values}
\item{flag}{a flag indicating nleqslv is converged or not, its value is equal to termcd of nleqslv}
\item{estim}{a list with two components named \code{coef} and \code{vcov}
containing information on the used estimators for both.}
\item{vcov}{a numeric matrix with the variance-covariance matrix for
the coefficients}
}
}
\description{
This is the basic computing engine called by \code{\link{gwfm}} to get the estimates for the coefficients and the variance-
covariance matrices. This function currently only spits out these components using the sandwich estimators.
}
\section{NOTE}{
 This function is mainly quoted from \code{\link{pim}} package by Joris Meys et al.(2020)
(https://cran.r-project.org/web/packages/pim/index.html)
}

\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
