% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_weight.R
\name{ipcw_weight}
\alias{ipcw_weight}
\title{Calculate IPCW weights in score function and sandwich variance estimation for a generalized win fraction regression model}
\usage{
ipcw_weight(
  tdata,
  trt_con,
  censor_cox,
  win_status,
  priority,
  determined_flag,
  L,
  n_ep,
  t_val
)
}
\arguments{
\item{tdata}{A data frame includes survival time, censoring indicator and related covariates for censoring}

\item{trt_con}{Given n subjects, there are n(n-1)/2 pairs for non-marginal structure and n(n-1) pairs for marginal structure, each row represents a pair. The analysis dataset trt_con contains the following variables:
\describe{
\item{pid_trt}{A vector for the subject id of the individuals from the "treatment" group}
\item{Delta_j_trt}{A vector for the event status of the jth endpoint (1=event, 0=censored) for the individuals from the "treatment" group}
\item{Y_j_trt}{A vector for the outcome of the jth endpoint for the individuals from the "treatment" group}
\item{pid_con}{A vector for the subject id of the individuals from the "control" group}
\item{Delta_j_con}{A vector for the event status of the jth endpoint (1=event, 0=censored) for the individuals from the "control" group}
\item{Y_j_con}{A vector for the outcome of the jth endpoint for the individuals from the "control" group}
}}

\item{censor_cox}{a list results from \code{\link{censoring_cox_est}}}

\item{win_status}{A data.frame for the win status of each pair for each endpoint, which is from \code{\link{win_strategy}}}

\item{priority}{Priority order (from the most to the least important). For example, given three endpoints with the importance order as Endpoint 1, Endpoint 2, and Endpoint 3, input priority = c(1,2,3).}

\item{determined_flag}{a value to indicate pairwise comparison is determined or not (1 refers to the comparison is determined; 0 refers to a tied comparison finally)}

\item{L}{restriction event time}

\item{n_ep}{the length of multiple endpoints}

\item{t_val}{truncation value for IPCW weights, which is a small value users can set}
}
\value{
A list with the IPCW weights used in score function and variance estimation
}
\description{
This function can compute IPCW weights used in score function and variance estimation for
a generalized win fraction regression model. Their detailed expressions are shown in Section 2.3 and Section 3 of paper (Cao et al., 2026).
}
\references{
Cao Z., Fang X. and Li F. Generalized win fraction regression for composite survival endpoints.
under review. 2026;0(0):1-35.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
