% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_csd.R
\name{merge_csd}
\alias{merge_csd}
\title{Merge composite survival data}
\usage{
merge_csd(datap, model, n_ep)
}
\arguments{
\item{datap}{a data frame includes multiple endpoints with corresponding censored indicators information for each subject.}

\item{model}{a single character value with possible values "difference" (the default), "marginal", "regular" or "customized". If the formula indicates a customized model (by the use of L() or R()), this parameter is set automatically to "customized".}

\item{n_ep}{the length of multiple endpoints.}
}
\value{
A data.frame for the information of id, all endpoints and corresponding censored indicator for each pair
}
\description{
An function to merge a pair (assume one from "treatment" group and the one from "control" group) composite survival data into a data frame, in which one row includes id, jth endpoint
and corresponding event status (1=event, 0=censored). Specifically, this function provides n(n-1)/2 pairs for non-marginal structure and n(n-1) pairs for marginal structure.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
