% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.gwfm.R
\name{model.matrix.gwfm}
\alias{model.matrix.gwfm}
\title{Create a model matrix for a generalized win fraction regression model}
\usage{
\method{model.matrix}{gwfm}(
  object,
  data,
  model = c("difference", "marginal", "customized"),
  ...
)
}
\arguments{
\item{object}{a gwfm.formula object that contains the formula necessary for constructing the model matrix.}

\item{data}{an optional argument specifying the data frame for which the model matrix should be constructed.}

\item{model}{a single character value with possible values "difference"
(the default), "marginal" or "customized". See also \code{\link{gwfm}}.}

\item{...}{extra arguments passed to or from other methods.
This is currently only implemented in concordance with the generic \code{\link[stats]{model.matrix}} function.}
}
\value{
a design matrix for a gwfm model
}
\description{
This function creates a model matrix for use in a generalized win fraction regression model (gwfm).
This model matrix can be passed to \code{\link{gwfm_fit}}.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
