% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gwfm_identity.R
\name{sim_gwfm_identity}
\alias{sim_gwfm_identity}
\title{Generate a simulated data for generalized win fraction regression model with identity link}
\usage{
sim_gwfm_identity(n, beta_d, beta_h, lambda_d, lambda_h, betac)
}
\arguments{
\item{n}{sample size in the simulated data}

\item{beta_d}{true regression coefficient in linear predictors for the first endpoint}

\item{beta_h}{true regression coefficient in linear predictors for the second endpoint}

\item{lambda_d}{shape parameter in exponential distribution for the first endpoint}

\item{lambda_h}{shape parameter in exponential distribution for the second endpoint}

\item{betac}{true regression coefficients in Cox model for censoring}
}
\value{
A data frame including id, y1(death time), y2(hospitalization time),
delta1(censored indicator for y1), delta2(censored indicator for y2),
x(covariate used in gwrm) and z(covariate used in censoring model)
}
\description{
This function can generated a simulated data for generalized win fraction regression model with identity link,
and two composite survival outcomes are generated from exponential distributions with different shape
parameters, and then modified through linear predictors with different regression coefficients
}
\examples{
n = 200  
beta_d = 10 
beta_h = 8
lambda_d = 0.15
lambda_h = 0.3 
betac = -5.7;  
mydata = sim_gwfm_identity(n,beta_d,beta_h,lambda_d,lambda_h,betac)

}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
