% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gwfm_logit.R
\name{sim_gwfm_logit}
\alias{sim_gwfm_logit}
\title{Generate a simulated data for generalized win fraction regression model with logit link}
\usage{
sim_gwfm_logit(n, beta, betac, lambda_d, lambda_h, alpha)
}
\arguments{
\item{n}{sample size in the simulated data}

\item{beta}{true regression coefficients in gwfrm model}

\item{betac}{true regression coefficients in Cox model for censoring}

\item{lambda_d}{baseline hazard in the marginal Cox model for the death time}

\item{lambda_h}{baseline hazard in the marginal Cox model for the hospitalization time}

\item{alpha}{Kendall’s tau between hospitalization time and death time}
}
\value{
A data frame including id, y1(death time), y2(hospitalization time),
delta1(censored indicator for y1), delta2(censored indicator for y2),
x1 and x2(covariates)
}
\description{
This function can generated a simulated data for generalized win fraction regression model with logit link,
and two composite survival outcomes are generated from a conditional Gumbel-Hougaard
copula model with Cox proportional hazards models as two marginal models
}
\examples{
n = 200  
beta = c(-0.5,0.5) 
betac = c(-1,0.5) 
lambda_d = 0.2
lambda_h = 2
alpha = 2
mydata = sim_gwfm_logit(n,beta,betac,lambda_d,lambda_h,alpha)

}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
