% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gwfm_probit.R
\name{sim_gwfm_probit}
\alias{sim_gwfm_probit}
\title{Generate a simulated data for generalized win fraction regression model with probit link}
\usage{
sim_gwfm_probit(n, alpha, betac, u, sig)
}
\arguments{
\item{n}{sample size in the simulated data}

\item{alpha}{true regression coefficients in marginal model}

\item{betac}{true regression coefficients in Cox model for censoring}

\item{u}{intercept term in both marginal models}

\item{sig}{standard error used in Gaussian copula model}
}
\value{
A data frame including id, y1(death time), y2(hospitalization time),
delta1(censored indicator for y1), delta2(censored indicator for y2),
x1 and x2(covariates)
}
\description{
This function can generated a simulated data for generalized win fraction regression model with probit link,
and two composite survival outcomes are generated from a conditional Gaussian copula model.
}
\examples{
n = 200  
alpha = 1
betac = -0.79
u = 6
sig = 1
set.seed(123456)
mydata = sim_gwfm_probit(n,alpha,betac,u,sig)

}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
