% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_strategy.R
\name{win_strategy}
\alias{win_strategy}
\title{The win strategy function}
\usage{
win_strategy(trt_con, priority)
}
\arguments{
\item{trt_con}{Given n subjects, there are n(n-1)/2 pairs for non-marginal structure and n(n-1) pairs for marginal structure, each row represents a pair. The analysis dataset trt_con contains the following variables:
\describe{
\item{pid_trt}{A vector for the subject id of the individuals from the "treatment" group}
\item{Delta_j_trt}{A vector for the event status of the jth endpoint (1=event, 0=censored) for the individuals from the "treatment" group}
\item{Y_j_trt}{A vector for the outcome of the jth endpoint for the individuals from the "treatment" group}
\item{pid_con}{A vector for the subject id of the individuals from the "control" group}
\item{Delta_j_con}{A vector for the event status of the jth endpoint (1=event, 0=censored) for the individuals from the "control" group}
\item{Y_j_con}{A vector for the outcome of the jth endpoint for the individuals from the "control" group}
}}

\item{priority}{Priority order (from the most to the least important). For example, given three endpoints with the importance order as Endpoint 1, Endpoint 2, and Endpoint 3, input priority = c(1,2,3).}
}
\value{
A data.frame for the win status of each pair for each endpoint.
}
\description{
An function to determine the win status for each pair based on the win strategy of Dong et al.(2022).
Specifically, one compares each subject in the "treatment" group with every subject in the "control"
group to determine the win status.
}
\references{
Cui, Y. and Huang, B. Package "WINS", CRAN R Repositary, 2025.

Dong, G., Huang, B., Wang, D., Verbeeck, J., Wang, J. and Hoaglin, D. Adjusting
win statistics for dependent censoring. Pharm Stat, 2021, 20(3):440--450.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}
}
